model/requirements.txt
fastapi
uvicorn
joblib
numpy
pandas
scikit-learn
requests
beautifulsoup4
tldextract
python-whois
validators
notebook
jupyterlab
ipykernel
matplotlib
seaborn
